Add the Version inner class to Lighthouse::Ticket.
This is to make it possible to YAML.load a .to_yaml'ed ticket in a
new session.  If no ticket has been loaded from Lighthouse yet, then
the Lighthouse::Ticket::Version class doesn't exist, and YAML.load
treats the Version objects as bags of attributes.  Adding the Version
inner class means it's always available, and we can use the #user_id
and similar accessors, even on loaded objects.

There may be other classes that need the same treatment, but the Version
inner class is extremely important, so it's the first one I ran into.